+++ RecipeChoice/WebContent/views/tip/knowledgeDetail.jsp
@@ -15,14 +15,15 @@
         }
         .wrap{
             width:1000px;
-            height:2500px;	/****자동으로 페이지 높이 조정하는 방법,,,,,,,*****/
+            overflow: hidden;
+            height: auto;	/****자동으로 페이지 높이 조정하는 방법,,,,,,,*****/
             margin: auto;
         }
         #header{height:150px;} 
         #content{height:auto;}
         #footer{height:10%;}
         #content>div{
-            height:100%;
+            height:auto;
             float:left;
         }
         #content_1{ width: 5%;  }
@@ -94,10 +95,10 @@
                     <p align="right">조회수 : <%=t.getTipCount() %></p>
                 </div>
 
-                <br><br><br>
+                <br>
                 <div class="detailArea">
                     <div align="center">
-                        <img src="resources/knowledge_upfiles/<%=t.getTipPicture() %>" width="400" height="250">
+                        <img src="resources/knowledge_upfiles/<%= t.getTipPicture() %>" width="400" height="250">
                     </div>
                     <br><br>
                     
@@ -108,46 +109,33 @@
 	                    <hr width="500px">
 	                    <div align="center" style="width: 400px;">
 	                       <%=t.getTipInfo() %>
-				                        간단설명간단설명간단설명간단설명간단설명간단설명간단설명
-				                        간단설명간단설명간단설명간단설명간단설명간단설명간단설명간단설명간단설명간단설명간단설명
 	                    </div>
 	                    <hr width="500px">
                     </div>
-                    <br><br><br>
 
                     <div id="detailContent" align="center" >
-                        <p id="conTitle" style="font-size: 16px; font-weight: bolder;">asdfasdfasdf</p>
-                        <img src="" width="400" height="250" style="margin-bottom: 20px;">
                         <p id="conInfo" style="word-break: break-all; width: 600px; margin-bottom: 100px;">
-                            asdfadfasdfaasdfasdfasdfadfsdfasd
-                            fasdfasdfadfasdfaasdfasdfasdfadfsdfasdfa
-                            sdfasdfadfasdfaasdfasdfasdfadfsdfasdfasdfasdfadfasdfaa
-                            sdfasdfasdfadfsdfasdfasdf
-                        </p>
-                        <p id="conTitle" style="font-size: 16px; font-weight: bolder;">asdfasdfasdf</p>
-                        <img src="" width="400" height="250" style="margin-bottom: 20px;">
-                        <p id="conInfo" style="word-break: break-all; width: 600px;">
-                            asdfadfasdfaasdfasdfasdfadfsdfasd
-                            fasdfasdfadfasdfaasdfasdfasdfadfsdfasdfa
-                            sdfasdfadfasdfaasdfasdfasdfadfsdfasdfasdfasdfadfasdfaa
-                            sdfasdfasdfadfsdfasdfasdf
+							<%=t.getTipContent()%>
                         </p>
                     </div>
                    
                 </div>
                 
-                <br><br>
+                <br>
                 <hr>
 
-                <%-- 관리자 일때만 보이게 --%>
                 <div align="right" style="margin-right: 50px;">
-                	<% if(t.getStatus().equals("N")){ %>
-                    	<a href="<%=contextPath%>/changeTipStatus.mn?tno=<%=t.getTipNo()%>" class="btn btn-danger">활성화</a>
-					<% }else{ %>
-                    	<a href="<%=contextPath%>/deleteKnowledge.mn?tno=<%=t.getTipNo()%>" class="btn btn-danger">삭제</a>
-                    <% } %>
-                    <a href="<%=contextPath%>/updateFormKnowledge.mn?tno=<%=t.getTipNo()%>" class="btn btn-success">수정</a>
-                    <a href="<%=contextPath%>/knowledgeList.mn" class="btn btn-secondary">목록</a>
+                	<% if(loginUser.getMemId().equals("admin")){ %>
+	                	<% if(t.getStatus().equals("N")){ %>
+	                    	<a href="<%=contextPath%>/changeTipStatus.mn?tno=<%=t.getTipNo()%>" class="btn btn-danger">활성화</a>
+						<% }else{ %>
+	                    	<a href="<%=contextPath%>/deleteKnowledge.mn?tno=<%=t.getTipNo()%>" class="btn btn-danger">삭제</a>
+	                    <% } %>
+	                    <a href="<%=contextPath%>/updateFormKnowledge.mn?tno=<%=t.getTipNo()%>" class="btn btn-success">수정</a>
+	                    <a href="<%=contextPath%>/knowledgeList.mn" class="btn btn-secondary">목록</a>
+	                <%}else{ %>
+	                    <a href="<%=contextPath%>/knowledgeList.tip?page=1" class="btn btn-secondary">목록</a>
+					<%} %>	                
                 </div>
                 
                 
@@ -155,6 +143,7 @@
         </div>
 
         </div>
+        <br><br>
 	 
 	 <%@ include file="../common/footerTemp.jsp" %>
 </body>
